# Conflicts:
#	eco-core/core-plugin/src/main/kotlin/com/willfp/ecopets/EcoPetsPlugin.kt
#	eco-core/core-plugin/src/main/kotlin/com/willfp/ecopets/pets/Pet.kt
#	eco-core/core-plugin/src/main/kotlin/com/willfp/ecopets/pets/PetDisplay.kt
#	eco-core/core-plugin/src/main/kotlin/com/willfp/ecopets/pets/entity/PetEntity.kt

+++ eco-core/core-plugin/src/main/kotlin/com/willfp/ecopets/api/EcoPetsAPI.kt
@@ -102,7 +102,7 @@ interface EcoPetsAPI {
     fun getPetXPRequired(
         player: OfflinePlayer,
         pet: Pet
-    ): Int
+    ): Double
 
     /**
      * Get experience to the next level.
